now able to open bookmark modal and.. edit title add labels remove labels create a new label
